adding categories

+++ admin/manage-users.php
@@ -34,23 +34,25 @@ $users = mysqli_query($connection, $query);
                 ?>
             </p>
         </div>
-    <?php elseif(isset($_SESSION['delete-user'])) : // shows if delete user was not successful?>
+    <?php elseif (isset($_SESSION['delete-user'])) : // shows if delete user was not successful
+    ?>
         <div class="alert__message error container">
-        <p>
-            <?= $_SESSION['delete-user'];
-            unset($_SESSION['delete-user']);
-             ?>
-        </p>
-    </div>
-    <?php elseif(isset($_SESSION['delete-user-success'])) : // shows if delete user was successful?>
+            <p>
+                <?= $_SESSION['delete-user'];
+                unset($_SESSION['delete-user']);
+                ?>
+            </p>
+        </div>
+    <?php elseif (isset($_SESSION['delete-user-success'])) : // shows if delete user was successful
+    ?>
         <div class="alert__message success container">
-        <p>
-            <?= $_SESSION['delete-user-success'];
-            unset($_SESSION['delete-user-success']);
-             ?>
-        </p>
-    </div>
-       <?php endif ?>
+            <p>
+                <?= $_SESSION['delete-user-success'];
+                unset($_SESSION['delete-user-success']);
+                ?>
+            </p>
+        </div>
+    <?php endif ?>
     <div class="container dashboard__container">
         <button id="show__sidebar-btn" class="sidebar__toggle"><i class="uil uil-angle-right-b"></i></button>
         <button id="hide__sidebar-btn" class="sidebar__toggle"><i class="uil uil-angle-left-b"></i></button>
@@ -89,29 +91,37 @@ $users = mysqli_query($connection, $query);
         </aside>
         <main>
             <h2>Manage Users</h2>
-            <table>
-                <thead>
-                    <tr>
-                        <th>Name</th>
-                        <th>Username</th>
-                        <th>Edit</th>
-                        <th>Delete</th>
-                        <th>Admin</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <?php while ($user = mysqli_fetch_assoc($users)) : ?>
+            <?php if (mysqli_num_rows($users) > 0) : ?>
+                <table>
+                    <thead>
                         <tr>
-                            <td><?= "{$user['firstname']} {$user['lastname']}" ?></td>
-                            <td><?= $user['username'] ?></td>
-                            <td><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?= $user['id'] ?>" class="btn sm">Edit</a></td>
-                            <td><a href="<?= ROOT_URL ?>admin/delete-user.php?id=<?= $user['id'] ?>" class="btn sm danger">Delete</a></td>
-                            <td><?= $user['is_admin'] ? 'Yes' : 'No' ?></td>
+                            <th>Name</th>
+                            <th>Username</th>
+                            <th>Edit</th>
+                            <th>Delete</th>
+                            <th>Admin</th>
                         </tr>
-                    <?php endwhile ?>
-                </tbody>
-            </table>
+                    </thead>
+                    <tbody>
+                        <?php while ($user = mysqli_fetch_assoc($users)) : ?>
+                            <tr>
+                                <td><?= "{$user['firstname']} {$user['lastname']}" ?></td>
+                                <td><?= $user['username'] ?></td>
+                                <td><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?= $user['id'] ?>" class="btn sm">Edit</a></td>
+                                <td><a href="<?= ROOT_URL ?>admin/delete-user.php?id=<?= $user['id'] ?>" class="btn sm danger">Delete</a></td>
+                                <td><?= $user['is_admin'] ? 'Yes' : 'No' ?></td>
+                            </tr>
+                        <?php endwhile ?>
+                    </tbody>
+                </table>
+            <?php else : ?>
+                <div class="alert__message error">
+                     <?= "No users found" ?>
+                    </div>
+            <?php endif ?>
         </main>
+    </div>
+    </main>
     </div>
 
 </section>
